requirements.txt
asyncpg
python-dotenv
discord.py
pytz
requests
pycoingecko